+++ client_token_integration_test.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/braintree-go/braintree-go/testhelpers"
+	"github.com/buu700/braintree-go-tmp/testhelpers"
 )
 
 // This test will fail unless you set up your Braintree sandbox account correctly. See TESTING.md for details.