+++ credit_card_integration_test.go
@@ -10,7 +10,7 @@ import (
 
 	"time"
 
-	"github.com/braintree-go/braintree-go/testhelpers"
+	"github.com/buu700/braintree-go-tmp/testhelpers"
 )
 
 func TestCreditCard(t *testing.T) {