+++ customer_integration_test.go
@@ -11,7 +11,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/braintree-go/braintree-go/testhelpers"
+	"github.com/buu700/braintree-go-tmp/testhelpers"
 )
 
 // This test will fail unless you set up your Braintree sandbox account correctly. See TESTING.md for details.