+++ examples/subscription/server.go
@@ -24,7 +24,7 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/braintree-go/braintree-go"
+	"github.com/buu700/braintree-go-tmp"
 )
 
 func main() {