+++ examples/transaction_3ds/server.go
@@ -19,7 +19,7 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/braintree-go/braintree-go"
+	"github.com/buu700/braintree-go-tmp"
 )
 
 func getBT() *braintree.Braintree {