+++ init_test.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/braintree-go/braintree-go/testhelpers"
+	"github.com/buu700/braintree-go-tmp/testhelpers"
 )
 
 const (